feat: AttributeConverter와 Converter를 통합한 CombinedConverter 및 기존 Converters 통합
