+++ src/main/java/com/LetMeDoWith/LetMeDoWith/entity/member/MemberStatusHistory.java
@@ -1,8 +1,8 @@
 package com.LetMeDoWith.LetMeDoWith.entity.member;
 
-import com.LetMeDoWith.LetMeDoWith.enums.attributeConverter.member.MemberStatusConverter;
-import com.LetMeDoWith.LetMeDoWith.enums.member.MemberStatus;
 import com.LetMeDoWith.LetMeDoWith.entity.BaseAuditEntity;
+import com.LetMeDoWith.LetMeDoWith.enums.converter.member.MemberStatusConverter;
+import com.LetMeDoWith.LetMeDoWith.enums.member.MemberStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
