+++ src/main/java/com/LetMeDoWith/LetMeDoWith/enums/converter/member/FollowTypeConverter.java
@@ -1,17 +1,16 @@
 package com.LetMeDoWith.LetMeDoWith.enums.converter.member;
 
-import org.springframework.core.convert.converter.Converter;
-
-import com.LetMeDoWith.LetMeDoWith.enums.common.FailResponseStatus;
-import com.LetMeDoWith.LetMeDoWith.enums.converter.AbstractConverter;
+import com.LetMeDoWith.LetMeDoWith.enums.converter.AbstractCombinedConverter;
 import com.LetMeDoWith.LetMeDoWith.enums.member.FollowType;
-import com.LetMeDoWith.LetMeDoWith.exception.RestApiException;
-import com.LetMeDoWith.LetMeDoWith.util.EnumUtil;
+import jakarta.persistence.Converter;
+import org.springframework.stereotype.Component;
 
-public class FollowTypeConverter extends AbstractConverter<FollowType> {
-
+@Converter(autoApply = true)
+@Component
+public class FollowTypeConverter extends AbstractCombinedConverter<FollowType> {
+	
 	public FollowTypeConverter() {
 		super(FollowType.class);
 	}
-
+	
 }
