+++ src/main/java/com/LetMeDoWith/LetMeDoWith/enums/converter/member/MemberStatusConverter.java
@@ -1,15 +1,16 @@
 package com.LetMeDoWith.LetMeDoWith.enums.converter.member;
 
-import com.LetMeDoWith.LetMeDoWith.enums.SocialProvider;
 import com.LetMeDoWith.LetMeDoWith.enums.converter.AbstractCombinedConverter;
+import com.LetMeDoWith.LetMeDoWith.enums.member.MemberStatus;
 import jakarta.persistence.Converter;
 import org.springframework.stereotype.Component;
 
 @Converter(autoApply = true)
 @Component
-public class SocialProviderConverter extends AbstractCombinedConverter<SocialProvider> {
+public class MemberStatusConverter extends AbstractCombinedConverter<MemberStatus> {
     
-    public SocialProviderConverter() {
-        super(SocialProvider.class);
+    public MemberStatusConverter() {
+        super(MemberStatus.class);
     }
+    
 }
